+++ app/src/main/java/com/example/adobongkangkong/domain/recipes/RecipeModels.kt
@@ -4,8 +4,8 @@ data class Recipe(
     val id: Long,
     val name: String,
     val ingredients: List<RecipeIngredient>,
-    val servingsYield: Int?,       // UX hint; required for per-serving + weight logging
-    val totalYieldGrams: Int?      // cooked final weight; required for per-cooked-gram
+    val servingsYield: Double?,       // UX hint; required for per-serving + weight logging
+    val totalYieldGrams: Double?      // cooked final weight; required for per-cooked-gram
 )
 
 data class RecipeIngredient(
